Move utils from test to main
