+++ tests/jballerina-debugger-integration-test/src/main/java/org/ballerinalang/debugger/main/utils/FileUtils.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.ballerinalang.debugger.test.utils;
+package org.ballerinalang.debugger.main.utils;
 
 import java.io.IOException;
 import java.nio.file.Files;
