+++ tests/jballerina-debugger-integration-test/src/main/java/org/ballerinalang/debugger/main/utils/client/connection/TestProcessStreamConnectionProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ballerinalang.debugger.test.utils.client.connection;
+package org.ballerinalang.debugger.main.utils.client.connection;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
