+++ tests/jballerina-debugger-integration-test/src/test/java/org/ballerinalang/debugger/test/adapter/run/SingleBalFileRunDebugTest.java
@@ -19,9 +19,9 @@
 package org.ballerinalang.debugger.test.adapter.run;
 
 import org.apache.commons.lang3.tuple.Pair;
+import org.ballerinalang.debugger.main.utils.BallerinaTestDebugPoint;
+import org.ballerinalang.debugger.main.utils.DebugUtils;
 import org.ballerinalang.debugger.test.DebugAdapterBaseTestCase;
-import org.ballerinalang.debugger.test.utils.BallerinaTestDebugPoint;
-import org.ballerinalang.debugger.test.utils.DebugUtils;
 import org.ballerinalang.test.context.BallerinaTestException;
 import org.eclipse.lsp4j.debug.StoppedEventArguments;
 import org.testng.Assert;
