+++ tests/jballerina-debugger-integration-test/src/test/java/org/ballerinalang/debugger/test/remote/BallerinaBuildRemoteDebugTest.java
@@ -26,7 +26,7 @@ import org.testng.annotations.Test;
 
 import java.io.File;
 
-import static org.ballerinalang.debugger.test.utils.DebugUtils.findFreePort;
+import static org.ballerinalang.debugger.main.utils.DebugUtils.findFreePort;
 
 /**
  * Test class to test positive and negative scenarios of remote debugging ballerina build command.
